split document cluster utils into manager and utils
